+++ examples/plot_3d_and_4d_niimg.py
@@ -82,6 +82,26 @@ for img in image.iter_img(rsn):
 
 
 ###############################################################################
+# Looping through selected volumes in a 4D file
+# ---------------------------------------------
+#
+# If we want to plot selected volumes in this 4D file, we can use index_img
+# with the `slice` constructor to select the desired volumes. 
+# 
+# Afterwards, we'll use iter_img to loop through them following the same 
+# formula as before.
+selected_volumes = image.index_img(rsn, slice(3, 5))
+
+###############################################################################
+# If you're new to Python, one thing to note is that the slice constructor
+# uses 0-based indexing. You can confirm this by matching these slices
+# to the previous plot above.
+
+for img in image.iter_img(selected_volumes):
+    plotting.plot_stat_map(img)
+
+
+###############################################################################
 # plotting.show is useful to force the display of figures when running
 # outside IPython
 plotting.show()