+++ nilearn/input_data/tests/test_nifti_labels_masker.py
@@ -6,6 +6,7 @@ test_masking.py and test_signal.py for details.
 """
 
 from nose.tools import assert_raises, assert_equal, assert_true
+import pytest
 import numpy as np
 
 import nibabel
@@ -14,6 +15,7 @@ from nilearn.input_data.nifti_labels_masker import NiftiLabelsMasker
 from nilearn._utils import testing, as_ndarray, data_gen
 from nilearn._utils.exceptions import DimensionError
 from nilearn._utils.testing import assert_less
+from nilearn.image import get_data
 
 
 def generate_random_img(shape, length=1, affine=np.eye(4),
@@ -113,7 +115,7 @@ def test_nifti_labels_masker_with_nans_and_infs():
                                                    affine=np.eye(4),
                                                    n_regions=n_regions)
     # nans
-    mask_data = mask_img.get_data()
+    mask_data = get_data(mask_img)
     mask_data[:, :, 7] = np.nan
     mask_data[:, :, 4] = np.inf
     mask_img = nibabel.Nifti1Image(mask_data, np.eye(4))
@@ -122,6 +124,49 @@ def test_nifti_labels_masker_with_nans_and_infs():
     sig = masker.fit_transform(fmri_img)
     assert_equal(sig.shape, (length, n_regions))
     assert_true(np.all(np.isfinite(sig)))
+
+
+def test_nifti_labels_masker_reduction_strategies():
+    """Tests:
+    1. whether the usage of different reduction strategies work.
+    2. whether unrecognised strategies raise a ValueError
+    3. whether the default option is backwards compatible (calls "mean")
+    """
+    test_values = [-2., -1., 0., 1., 2]
+
+    img_data = np.array([[test_values,
+                          test_values]])
+
+    labels_data = np.array([[[0, 0, 0, 0, 0],
+                             [1, 1, 1, 1, 1]]], dtype=np.int8)
+
+    affine = np.eye(4)
+    img = nibabel.Nifti1Image(img_data, affine)
+    labels = nibabel.Nifti1Image(labels_data, affine)
+
+    # What NiftiLabelsMasker should return for each reduction strategy?
+    expected_results = {"mean": np.mean(test_values),
+                        "median": np.median(test_values),
+                        "sum": np.sum(test_values),
+                        "minimum": np.min(test_values),
+                        "maximum": np.max(test_values),
+                        "standard_deviation": np.std(test_values),
+                        "variance": np.var(test_values)}
+
+    for strategy, expected_result in expected_results.items():
+        masker = NiftiLabelsMasker(labels, strategy=strategy)
+        # Here passing [img] within a list because it's a 3D object.
+        result = masker.fit_transform([img]).squeeze()
+        assert result == expected_result
+
+    with pytest.raises(ValueError, match="Invalid strategy 'TESTRAISE'"):
+        NiftiLabelsMasker(
+            labels,
+            strategy="TESTRAISE"
+        )
+
+    default_masker = NiftiLabelsMasker(labels)
+    assert default_masker.strategy == "mean"
 
 
 def test_nifti_labels_masker_resampling():
@@ -206,7 +251,7 @@ def test_nifti_labels_masker_resampling():
                                    masker.labels_img_.affine)
     assert_equal(masker.mask_img_.shape, masker.labels_img_.shape[:3])
 
-    uniq_labels = np.unique(masker.labels_img_.get_data())
+    uniq_labels = np.unique(get_data(masker.labels_img_))
     assert_equal(uniq_labels[0], 0)
     assert_equal(len(uniq_labels) - 1, n_regions)
 
@@ -254,7 +299,7 @@ def test_nifti_labels_masker_resampling():
                                    resampling_target=resampling_target)
         transformed = masker.fit_transform(fmri_img)
         resampled_labels_img = masker._resampled_labels_img_
-        n_resampled_labels = len(np.unique(resampled_labels_img.get_data()))
+        n_resampled_labels = len(np.unique(get_data(resampled_labels_img)))
         assert_equal(n_resampled_labels - 1, transformed.shape[1])
         # inverse transform
         compressed_img = masker.inverse_transform(transformed)
@@ -264,8 +309,8 @@ def test_nifti_labels_masker_resampling():
         transformed2 = masker.fit_transform(fmri_img)
         # inverse transform again
         compressed_img2 = masker.inverse_transform(transformed2)
-        np.testing.assert_array_equal(compressed_img.get_data(),
-                                      compressed_img2.get_data())
+        np.testing.assert_array_equal(get_data(compressed_img),
+                                      get_data(compressed_img2))
 
 
 def test_standardization():