+++ nilearn/plotting/displays.py
@@ -29,6 +29,7 @@ from .. import _utils
 from ..image import new_img_like
 from ..image.resampling import (get_bounds, reorder_img, coord_transform,
                                 get_mask_bounds)
+from nilearn.image import get_data
 
 
 ###############################################################################
@@ -403,7 +404,7 @@ class GlassBrainAxes(BaseAxes):
                 if self.direction == 'r' and xc >= 0:
                     relevant_coords.append(cidx)
                 elif self.direction == 'l' and xc <= 0:
-                        relevant_coords.append(cidx)
+                    relevant_coords.append(cidx)
             xdata = xdata[relevant_coords]
             ydata = ydata[relevant_coords]
             # if marker_color is string for example 'red' or 'blue', then
@@ -414,6 +415,9 @@ class GlassBrainAxes(BaseAxes):
             if not isinstance(marker_color, _utils.compat._basestring) and \
                     len(marker_color) != 1:
                 marker_color = marker_color[relevant_coords]
+
+            if not isinstance(marker_size, numbers.Number):
+                marker_size = np.asarray(marker_size)[relevant_coords]
 
         defaults = {'marker': 'o',
                     'zorder': 1000}
@@ -889,7 +893,7 @@ class BaseSlicer(object):
             The color used to display the edge map
         """
         img = reorder_img(img, resample='continuous')
-        data = img.get_data()
+        data = get_data(img)
         affine = img.affine
         single_color_cmap = colors.ListedColormap([color])
         data_bounds = get_bounds(data.shape, img.affine)